requirements.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
lightning-utilities==0.15.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
packaging==25.0
pandas==2.3.2
pandas-stubs==2.3.2.250827
pillow==11.3.0
propcache==0.3.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.6.0
torchmetrics==1.8.2
torchvision==0.21.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
